refactor(core): new utility function to add state
